pyproject.toml
[tool.poetry]
name = "bird-arcwise"
version = "0.0.0"
description = "BIRD-SQL submission"
authors = ["Arcwise Team <arcwise-team@arcwise.app>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.8.2"
sqlglot = "^25.6.0"
litellm = "^1.41.23"
python-dotenv = "^1.0.1"
tqdm = "^4.66.4"
click = "^8.1.7"


[tool.ruff]
line-length = 100


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
